refactor(addon): move ColorFormatSelector out of switcher and into the addon
Color format is a blocks-rendering concern (hex / rgb / hsl / oklch /
raw — the format swatchbook blocks use to stringify sub-values) not a
theming one. Keeping it in the shared switcher made the docs-site
navbar carry a dropdown with no consumer. The switcher's API contract
is "framework-agnostic theme switcher UI"; this selector belongs to
the addon.

Move the component into packages/addon/src/ColorFormatSelector.tsx,
strip the SwitcherColorFormat type + ColorFormatSelector export from
the switcher package. Addon's manager.tsx imports the local file and
still slots it into the ThemeSwitcher's `footer` prop so the toolbar
popover layout is unchanged.

The addon-local component reuses the `sb-switcher__*` class names
since the switcher's CSS is already loaded whenever the popover
renders — no CSS duplication.
