+++ packages/switcher/src/index.ts
@@ -1,3 +1,2 @@
-export { ColorFormatSelector, type ColorFormatSelectorProps } from '#/ColorFormatSelector.tsx';
 export { ThemeSwitcher, type ThemeSwitcherProps } from '#/ThemeSwitcher.tsx';
-export type { SwitcherAxis, SwitcherColorFormat, SwitcherPreset, SwitcherTheme } from '#/types.ts';
+export type { SwitcherAxis, SwitcherPreset, SwitcherTheme } from '#/types.ts';
